Плитки окна проекта — лестницей, и колесо с Ctrl её двигает
Размер плиток правой колонки знал три положения — список, 64 и 128 — и менялся
только ползунком. Промежуточных не было не по недосмотру: только на 64 и 128
силуэт сетки 16 ложится в пиксели на каждом масштабе с шагом в четверть, у 48 и
96 при 125 и 175 % край размывался целиком.

Теперь силуэт на промежуточных ступенях сажает на пиксели сам AxIcon (подмодуль
значков): каждая вершина — к ближайшему пикселю значка, сдвиг меньше полупикселя,
прямые края на границах пикселей. Кайма папки у 48 при 125, 150 и 175 % упала со
134, 108 и 189 пикселей до 8, 9 и 18. Тема получила малую ступень
AxTileGlyphSizeSmall 32 и шаг AxTileGlyphSizeStep 16, и ступеней стало семь,
32…128; ширина плитки идёт по прямой через две точки темы, подложка не крупнее
обычной ступени. Размеры ступени колонка ставит списку (TileMetrics), разметка
берёт их у него, а IconPixelTests принимает эту привязку единственным способом
растянуть значок и мерит кадром каждую ступень на каждом масштабе. SDK 7.2.

Колесо с Ctrl над колонкой — ступень за щелчок, как в проводнике и Unity: мельче
малой — список, колонка при этом не листается; доли тачпада копятся до целого
щелчка. С клавиатуры — Ctrl с плюсом и минусом, Ctrl+0 — обычная ступень;
подсказка ползунка называет все три дороги. Выбранная плитка остаётся в виду.
В настройке теперь размер в точках, а не номер ступени: номер съезжал бы с каждой
правкой лестницы. Прежние 1 и 2 читаются как обычные и крупные плитки.

Живая проверка нашла два огреха, и оба исправлены: Ctrl+минус из плитки в список
терял фокус — клавиатура теперь возвращается на тот же предмет, — а на малой
ступени две строки подписи ломали имена посреди слова, и мельче обычной ступени
подпись идёт в строку с многоточием. Саботажем проверено девятнадцать мест в
трёх репозиториях, все пойманы.

1475 тестов, 0 предупреждений. Запись 240 в плане.

+++ src/Modules/ArxisStudio.Modules.Project/Panels/ProjectPanel.cs
@@ -501,8 +501,8 @@ public sealed class ProjectPanel : ToolWindow
             _view.Query.Text = string.Empty;
     }
 
-    /// <summary>Ползунок сдвинули — ступень уходит в настройки.</summary>
-    private void Resize(int size)
+    /// <summary>Ступень сменили — ползунком, колесом или с клавиатуры, — и её размер уходит в настройки.</summary>
+    private void Resize(double size)
     {
         if (ProjectSettings.Read(Context.Settings).IconSize != size)
             Context.Settings.Set(ProjectSettings.IconSizeKey, size);
